Add change color action to EditNoteActivity
